requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
openai==1.12.0
anthropic==0.19.0
kafka-python==2.0.2
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
sentence-transformers==2.3.1
numpy==1.26.3
grpcio==1.62.0
grpcio-tools==1.62.0
python-dotenv==1.0.0
pydantic==2.5.3